# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Process-Sync-Api",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
        "/deployments": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "list deployments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployment"
                ],
                "summary": "list deployments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "deployment id",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "default id.asc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "add the fields 'diagram', 'definition_id' and 'error' to the results",
                        "name": "extended",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma seperated list of network-ids used to filter the deployments",
                        "name": "network_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Deployment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/deployments/{networkId}": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "deploy process; prepared process may be requested from the process-fog-deployment service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployment"
                ],
                "summary": "deploy process",
                "parameters": [
                    {
                        "description": "deployment",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/deploymentmodel.Deployment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/deployments/{networkId}/{deploymentId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get deployment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployment"
                ],
                "summary": "get deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "deployment id",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Deployment"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "delete deployment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployment"
                ],
                "summary": "delete deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "deployment id",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/deployments/{networkId}/{deploymentId}/metadata": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get deployment metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployment",
                    "metadata"
                ],
                "summary": "get deployment metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "deployment id",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DeploymentMetadata"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/deployments/{networkId}/{deploymentId}/start": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "start deployed process; a process may expect parameters on start. these cna be passed as query parameters. swagger allows no arbitrary/dynamic parameter names, which means a query wit parameters must be executed manually",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployment"
                ],
                "summary": "start deployed process",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "deployment id",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "deploymentmodel.ConditionalEvent": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "qos": {
                    "type": "integer"
                },
                "script": {
                    "type": "string"
                },
                "selection": {
                    "$ref": "#/definitions/deploymentmodel.Selection"
                },
                "value_variable": {
                    "type": "string"
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "deploymentmodel.Deployment": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "diagram": {
                    "$ref": "#/definitions/deploymentmodel.Diagram"
                },
                "elements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/deploymentmodel.Element"
                    }
                },
                "executable": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "incident_handling": {
                    "$ref": "#/definitions/deploymentmodel.IncidentHandling"
                },
                "name": {
                    "type": "string"
                },
                "start_parameter": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/deploymentmodel.ProcessStartParameter"
                    }
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "deploymentmodel.Diagram": {
            "type": "object",
            "properties": {
                "svg": {
                    "type": "string"
                },
                "xml_deployed": {
                    "type": "string"
                },
                "xml_raw": {
                    "type": "string"
                }
            }
        },
        "deploymentmodel.Element": {
            "type": "object",
            "properties": {
                "bpmn_id": {
                    "type": "string"
                },
                "conditional_event": {
                    "$ref": "#/definitions/deploymentmodel.ConditionalEvent"
                },
                "group": {
                    "type": "string"
                },
                "message_event": {
                    "$ref": "#/definitions/deploymentmodel.MessageEvent"
                },
                "name": {
                    "type": "string"
                },
                "notification": {
                    "$ref": "#/definitions/deploymentmodel.Notification"
                },
                "order": {
                    "type": "integer"
                },
                "task": {
                    "$ref": "#/definitions/deploymentmodel.Task"
                },
                "time_event": {
                    "$ref": "#/definitions/deploymentmodel.TimeEvent"
                }
            }
        },
...
        "deploymentmodel.IncidentHandling": {
            "type": "object",
            "properties": {
                "notify": {
                    "type": "boolean"
                },
                "restart": {
                    "type": "boolean"
                }
            }
        },
        "deploymentmodel.MessageEvent": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "flow_id": {
                    "type": "string"
                },
                "selection": {
                    "$ref": "#/definitions/deploymentmodel.Selection"
                },
                "use_marshaller": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "deploymentmodel.Notification": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "deploymentmodel.ProcessStartParameter": {
            "type": "object",
            "properties": {
                "default": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "properties": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "deploymentmodel.Task": {
            "type": "object",
            "properties": {
                "parameter": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "prefer_event": {
                    "type": "boolean"
                },
                "retries": {
                    "type": "integer"
                },
                "selection": {
                    "$ref": "#/definitions/deploymentmodel.Selection"
                }
            }
        },
        "deploymentmodel.TimeEvent": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_SENERGY-Platform_process-sync_pkg_model_camundamodel.Variable": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {},
                "valueInfo": {}
            }
        },
...
        "model.Deployment": {
            "type": "object",
            "properties": {
                "deploymentTime": {},
                "id": {
                    "type": "string"
                },
                "is_placeholder": {
                    "type": "boolean"
                },
                "marked_as_missing": {
                    "type": "boolean"
                },
                "marked_for_delete": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "network_id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "sync_date": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                }
            }
        },
        "model.DeploymentMetadata": {
            "type": "object",
            "properties": {
                "camunda_deployment_id": {
                    "type": "string"
                },
                "deployment_model": {
                    "$ref": "#/definitions/model.DeploymentWithEventDesc"
                },
                "is_placeholder": {
                    "type": "boolean"
                },
                "marked_as_missing": {
                    "type": "boolean"
                },
                "marked_for_delete": {
                    "type": "boolean"
                },
                "network_id": {
                    "type": "string"
                },
                "process_parameter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_SENERGY-Platform_process-sync_pkg_model_camundamodel.Variable"
                    }
                },
                "sync_date": {
                    "type": "string"
                }
            }
        },
        "model.DeploymentWithEventDesc": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "device_id_to_local_id": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "diagram": {
                    "$ref": "#/definitions/deploymentmodel.Diagram"
                },
                "elements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/deploymentmodel.Element"
                    }
                },
                "event_descriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EventDesc"
                    }
                },
                "executable": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "incident_handling": {
                    "$ref": "#/definitions/deploymentmodel.IncidentHandling"
                },
                "name": {
                    "type": "string"
                },
                "service_id_to_local_id": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "start_parameter": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/deploymentmodel.ProcessStartParameter"
                    }
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}